gtk: use f64 everywhere instead of i16/f16/c_int

+++ src/apprt/gtk/Surface.zig
@@ -96,7 +96,7 @@ pub const Container = union(enum) {
             }
         }
 
-        pub fn equalize(self: Elem) i16 {
+        pub fn equalize(self: Elem) f64 {
             return switch (self) {
                 .surface => 1,
                 .split => |s| s.equalize(),
